# More permissive Index typing