- `Index.isel`: more permissive return type (any Xarray Index)

This allows an index to return a new index of another type, e.g., a
1-dimensional CoordinateTransformIndex to return a PandasIndex when a
new transform cannot be computed (selection at arbitrary locations).

- `Index.equals`: more permissive `other` type (any Xarray Index)

Xarray alignment logic is such that Xarray indexes are always compared
with other indexes of the same type. However, this is not necessarily
the case for "meta" indexes (i.e., indexes encapsulating one or more
index objects that may have another type) that are dispatching `equals`
to their wrapped indexes.

More context and concrete examples in https://github.com/corteva/rioxarray/pull/846#discussion_r1965142497 and https://github.com/corteva/rioxarray/pull/846#discussion_r1964042312.