@hmaarrfk 2025-02-21
Thank you for tackling this in a very cool way!!!

On main:
```
software_timestamp.nc
h5netcdf  lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 16285.88it/s]
h5netcdf  computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 19672.38it/s]
netcdf4   lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 16254.36it/s]
netcdf4   computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 19496.47it/s]
software_timestamp_float64.nc
h5netcdf  lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 16260.72it/s]
h5netcdf  computed  : 100%|█████████████████| 100000/100000 [00:04<00:00, 20045.22it/s]
netcdf4   lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 16329.55it/s]
netcdf4   computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 19780.63it/s]
```

On this branch:

```
software_timestamp.nc
h5netcdf  lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 15767.95it/s]
h5netcdf  computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 19327.47it/s]
netcdf4   lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 15904.21it/s]
netcdf4   computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 18807.63it/s]
software_timestamp_float64.nc
h5netcdf  lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 15825.22it/s]
h5netcdf  computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 18786.63it/s]
netcdf4   lazy      : 100%|█████████████████| 100000/100000 [00:06<00:00, 15515.68it/s]
netcdf4   computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 19033.53it/s]
```

On v2024.03.0
```
software_timestamp.nc
h5netcdf  lazy      : 100%|█████████████████| 100000/100000 [00:09<00:00, 10237.03it/s]
h5netcdf  computed  : 100%|██████████████████| 100000/100000 [00:14<00:00, 6672.00it/s]
netcdf4   lazy      : 100%|██████████████████| 100000/100000 [00:10<00:00, 9718.41it/s]
netcdf4   computed  : 100%|██████████████████| 100000/100000 [00:14<00:00, 6923.85it/s]
software_timestamp_float64.nc
h5netcdf  lazy      : 100%|█████████████████| 100000/100000 [00:09<00:00, 10752.14it/s]
h5netcdf  computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 17995.04it/s]
netcdf4   lazy      : 100%|█████████████████| 100000/100000 [00:09<00:00, 10957.65it/s]
netcdf4   computed  : 100%|█████████████████| 100000/100000 [00:05<00:00, 17768.37it/s]
```

It seems that the performance improvements that were brought over 2024 are maintained (I think the 300its/s drop is real, but the main gains were in the 1000s of it/s in my benchmarks). In 2024 there was also a bug that made timestamps very slow to load in which is why my benchmark tests against data where the timestamp is also stored as a float64.

```python
import xarray
from tqdm import tqdm

for filename in [
    'software_timestamp.nc',
    'software_timestamp_float64.nc',
]:
    print(filename)
    for engine in ['h5netcdf', 'netcdf4']:
        for computed in [False, True]:
            dataset = xarray.open_dataset(filename, engine=engine)
            N_frames = dataset.sizes['frame_number']
            N_iters = 100_000
            software_timestamp = dataset['software_timestamp']
            if computed:
                software_timestamp = software_timestamp.compute()

            if computed:
                computed_str = "computed"
            else:
                computed_str = "lazy"
            desc = f"{engine:10s}{computed_str:10s}"
            for i in tqdm(range(N_iters), desc=desc):
                software_timestamp.isel(frame_number=i % N_frames)

            dataset.close()
```